SYS_GIFT 中的 rnd 可能是 Long 类型

+++ src/main/java/com/hiczp/bilibili/live/danmu/api/entity/SysGiftEntity.java
@@ -5,6 +5,6 @@ package com.hiczp.bilibili.live.danmu.api.entity;
  */
 public class SysGiftEntity extends JSONEntity {
     public String msg;
-    public Integer rnd;
+    public Long rnd;
     public Integer uid;
 }
